requirements.txt
Pillow==9.4.0
torch==1.10.0+cu102
torchvision==0.11.0+cu102
